update to geo_concerns 0.2.0
